Use Bundle type to load images, remove systemName
